% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_wind_vectors.R
\name{average_wind_vectors}
\alias{average_wind_vectors}
\title{Average wind vectors}
\usage{
average_wind_vectors(u_wind, v_wind, na.rm = FALSE)
}
\arguments{
\item{u_wind}{numeric; vector of values}

\item{v_wind}{numeric; vector of values}

\item{na.rm}{logical; remove NAs or not}
}
\description{
Averages wind vectors but retains average absolute wind speed
}
\examples{
average_wind_vectors(u_wind = c(2, -2), v_wind = c(0, 4))

}
