% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyancy_freq.R
\name{buoyancy_freq}
\alias{buoyancy_freq}
\title{Calculate buoyancy frequency}
\usage{
buoyancy_freq(wtr, depths, dz = 0.1, rLA_method = FALSE)
}
\arguments{
\item{wtr}{numeric; vector of water temperatures in degC}

\item{depths}{numeric; vector of depths in m, positive downward}

\item{dz}{numeric; depth spacing to interpolate to, defaults to 0.1 m}

\item{rLA_method}{logical; use rLakeAnalyzer method to calculate z, defaults to FALSE}
}
\description{
Calculation of the buoyancy frequency "N", in rad/s, returns data.frame
}
\examples{
buoyancy_freq(c(14, 13, 9, 8), c(0, 2, 4, 6))

}
