% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_V1V2_modes.R
\name{find_V1V2_modes}
\alias{find_V1V2_modes}
\title{Finds wave periods, thermocline depth, and metalimnion boundaries}
\usage{
find_V1V2_modes(wtr, depths, basin_length, rLA_method = FALSE)
}
\arguments{
\item{wtr}{numeric; vector of water temperatures in degC}

\item{depths}{numeric; vector of depths in m, positive downward}

\item{basin_length}{numeric; basin length in m}

\item{rLA_method}{logical; use rLakeAnalyzer's buoyancy frequency calculation,
defaults to FALSE}
}
\description{
Uses the temperature profile (z,T) and computes profile of buoyancy
frequency (zN,N) to solve for the V1 and V2 seiche modal structure. The
thermocline depth (zt) is defined as the depth of maximum displacement
during a V1 seiche and is the peak in the V1 modal structure. The
boundaries of the metalimnion (zm1, zm2) are the depths of maximum
displacement during a V2 seiche and are identified from the peaks in the
V2 modal structure. From this and the basin length L, periods of the V1H1
(T1) and V2H1 (T2) seiches (in s) are computed. This is based on Munnich et al
1992 and solves for the modes using many layers as opposed to a two or
three layer approximation.
Returns a list
}
\examples{
find_V1V2_modes(c(14, 13, 9, 8), c(0, 2, 4, 6), 7400)

}
