% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_speed_to_vectors.R
\name{wind_speed_to_vectors}
\alias{wind_speed_to_vectors}
\title{Wind speed and direction into vectors}
\usage{
wind_speed_to_vectors(speed, direction)
}
\arguments{
\item{speed}{numeric;}

\item{direction}{numeric; wind direction in degrees, 0 = north, 90 = east, etc}
}
\description{
Function to convert wind speed and direction into vectors, returns vector
}
\examples{
wind_speed_to_vectors(5, 45)

}
