% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_vectors_to_speed.R
\name{wind_vectors_to_speed}
\alias{wind_vectors_to_speed}
\title{Wind vectors to speed and dir}
\usage{
wind_vectors_to_speed(u_wind, v_wind)
}
\arguments{
\item{u_wind}{numeric;}

\item{v_wind}{numeric;}
}
\description{
Function to convert wind vectors into speed and direction in degrees, returns vector
Returned direction: 0 is north, 90 is east, etc
}
\examples{
wind_vectors_to_speed(-2, 3)

}
